requirements.txt
Pillow==8.4.0
praw==7.5.0
prawcore==2.3.0
DateTime==4.3
requests==2.26.0
